requirements.txt
openai>=1.3.0
python-dotenv
requests
beautifulsoup4
mysql-connector-python
